webapp(authority)/requirements.txt
beautifulsoup4==4.9.1
Django==2.2
django-bootstrap4==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
Pillow==7.2.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
